manifest: use private mime type database
This fixes #1221

The private mime database starts out as a copy of the global mime database.
In particular, it contains all well-known types like image/jpeg and image/png.
We add/overwrite our own types into that database instead of the global one.
This private database is immune to subsequent changes to the global mime
database, such as mimetypes.init() or conflicting add_type(...) calls.